Use async wrapper in test files, close camera after use
- Most of the tests use `async` calls via the wrapped `usb` device, wrapped in a mix of `sync` loops (usually `array.forEach()`) and "loose" promises (logging to `stdout`/`stderr`).
- Execution of test files does not finish if the camera object is not explicitly closed after use, as the `usb` device is still open and active.
- Because of the "loose" promises, there is no clear way to tell when the camera object is no longer in use.
- This commit fixes affected tests by adding a top-level async `run()` wrapper function, and converting `.forEach()` loops to `for...of` loops with `await`ed camera device calls inside.

The test files (like the rest of this project) uses a mix of outdated code styles. The upstream refactoring from v1 to v2 was never completed, and left in a messy state. All code should obviously be fixed "at some point", but since the upstream project has been dead for years I've avoided it in this fork. While I need this code to work for `uvcc` to work, fixing and cleaning up all (any!) of it more than necessary is not a priority. Patches welcome